Improved version of prev preload. Use landing component instead of window.
